// src/api/model/configUpdate.ts
/**
 * Generated by orval v7.7.0 🍺
 * Do not edit manually.
 * FinMan API
 * OpenAPI spec version: 1.0.0
 */
import type { ConfigUpdateTypeData } from "./configUpdateTypeData";
import type { ConfigUpdateValueData } from "./configUpdateValueData";
import type { ConfigUpdateJsonData } from "./configUpdateJsonData";
import type { ConfigUpdateAddValue } from "./configUpdateAddValue";

/**
 * Схема для оновлення конфігурації
 */
export interface ConfigUpdate {
  type_data?: ConfigUpdateTypeData;
  value_data?: ConfigUpdateValueData;
  json_data?: ConfigUpdateJsonData;
  add_value?: ConfigUpdateAddValue;
}